chatbot_V2/requirements.txt
# FastAPI Authentication API Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]>=2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.1
python-dotenv==1.0.0
email-validator==2.1.0